Fix up obsolete classes so that we can build the JARs

+++ trunk/CRAM-UI/src/uk/ac/lkl/cram/ui/obsolete/PresentationTableModel.java
@@ -124,10 +124,10 @@ public class PresentationTableModel extends AbstractTableModel {
                 mp.setFee((Integer) value);
                 break;
             case 3:
-                mp.setJuniorCost((Float) value);
+                mp.setJuniorCost((Integer) value);
                 break;
             case 4:
-                mp.setSeniorCost((Float) value);
+                mp.setSeniorCost((Integer) value);
                 break;
             default:
                 throw new RuntimeException("Invalid column: " + column);
